requirements.txt
albucore==0.0.24
albumentations==2.0.8
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.28.0
fonttools==4.62.1
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
narwhals==2.19.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pandas==3.0.2
pillow==12.2.0
protobuf==7.34.1
pyarrow==23.0.1
pydantic==2.13.1
pydantic_core==2.46.1
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
simsimd==6.5.16
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
